Drop DS requirement on runtime class
The monitor pods require a custom SELinux policy to be installed. The
monitor daemon set (DS) should thus be created after that, in order to
avoid polluting the logs with transient errors.

The current code base does this in a bit convoluted and slightly bogus
way :
1) it is assumed that the SELinux policy results from the creation of
   the runtime class
2) the DS creation code is located in a place where only an extra
   check allows to know about the runtime class existence

[1] is wrong since the SELinux policy resuls from the creation of
the security context constraints (SCC). As already suggested in
a comment, it seems that the creation of the DS should happen
just after the SCC is created to address [2].

This removes a user of KataConfigStatus::RuntimeClass that will be
renamed in the subsequent patch.

Signed-off-by: Greg Kurz <groug@kaod.org>