Rename RuntimeClass field to RuntimeClasses
PR #344 changed the type of KataConfigStatus::RuntimeClass from string
to array of string. This change in the CRD isn't supported and prevents
upgrades from an older operator if a KataConfig is present. The new CSV
will stay in the `Pending` state forever and the following error is
reported by the install plan :

message: 'error validating existing CRs against new CRD''s schema for "kataconfigs.kataconfiguration.openshift.io":
      error validating custom resource against new schema for KataConfig /example-kataconfig:
      [].status.runtimeClass: Invalid value: "string": status.runtimeClass in body
      must be of type array: "string"'

Simply rename the field to avoid the issue.

Signed-off-by: Greg Kurz <groug@kaod.org>