Always regenerate KataConfigStatus::KataNodes
The list of nodes where kata is deployed is part of the status of
the KataConfig. It is expected that the controller ensures this
list is up-to-date during all reconcile runs. Changes to this
list usually happen during installation/deinstallation of kata
on some nodes. But it is also possible that reconcile is called
and the status is lacking the list of nodes for some reason, e.g.
KataConfig was edited externally or KataConfig predates PR #329.

Stop restricting the regeneration of this list to only happen when
kata is being deployed or removed. This might result in some more
trafic between the client and server but idempotency should prevail.

Signed-off-by: Greg Kurz <groug@kaod.org>